Various fixes to get drivers functional (WIP)
This builds ontop of the last patch to get things functional. Includes:

    - Fix notifications layer to prepare events in their new format. Moves
      notificatin processing to the drivers themselves.

    - Repairs what used to be the TenantRouterCache and creates a new driver method
      for resolving resources by tenant. This fixes issues with current patch where
      events with no associated resource cause problems.

    - Converts Resource class to a namedtuple for now..

    - Avoids using a wrapper to update cache on the router driver--adds unnecessary
      complexity ATM

    - fixes some other method calls

    - Adds a driver method for listing ports of a logical resource

    - Adds get_state() driver method for reporting the state of the driver's
      resource. This is needed by the instance manager now that we've decoupled
       the instance state from logical resource state.

    - Finishes/fixes the router state sync'ing, adds required router driver
      method

    - Moves all usage of the akanda api client from instance manager to the router
      driver. ATM this client lib is router-specific.

    - Moves update_config, get_interfaces and is_alive to driver methods--
      different appliances will need to do these things in different ways
      (we currently use the akanda_client/router API service for them)

    - Moves router configuration to its own [router] group and deprecates
      the old options in [default]

This'll merge down to the patch below it when its ready.

+++ akanda/rug/instance_manager.py
@@ -13,13 +13,13 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
+
 from datetime import datetime
 import time
 
 from oslo_config import cfg
 
 from akanda.rug.drivers import states
-from akanda.rug.api import akanda_client
 from akanda.rug.common.i18n import _LE, _LI
 
 CONF = cfg.CONF
@@ -42,15 +42,11 @@ CONF.register_opts(INSTANCE_MANAGER_OPTS)
 
 
 def synchronize_driver_state(f):
-    """triggers a drivers synchronize_state function after executing something
-    that effects the state locally.
-
-    NOTE: This decorator requires the function it is wrapping to return a valid
-    state value.
-    """
+    """Wrapper that triggers a driver's synchronize_state function"""
     def wrapper(self, *args, **kw):
         state = f(self, *args, **kw)
-        return self.driver.synchronize_state(state)
+        self.driver.synchronize_state(*args, state=state)
+        return state
     return wrapper
 
 
@@ -71,17 +67,14 @@ def ensure_cache(f):
                 worker_context.nova_client.get_instance_info(self.driver.name)
             )
 
-            # if all that fails then we have to log and set the instance state
-            # to DOWN
-            if self.instance_info is None:
-                self.log.error(_LE('no backing instance, marking as down'))
-                self.state = states.DOWN
-                return
+            if self.instance_info:
+                (
+                    self.instance_info.management_port,
+                    self.instance_info.ports
+                ) = worker_context.neutron.get_ports_for_instance(
+                    self.instance_info.id_
+                )
 
-        if self.state == states.GONE:
-            self.log.info(_LI('not booting deleted instance'))
-            return
-        else:
             return f(self, worker_context, *args, **kw)
 
     return wrapper
@@ -142,18 +135,29 @@ class InstanceManager(object):
         """
         self._boot_counter.reset()
 
-    @ensure_cache
     @synchronize_driver_state
     def update_state(self, worker_context, silent=False):
-        """Updates state of logical resource
+        """Updates state of the instance and, by extension, its logical resource
 
         :param worker_context:
         :param silent:
         :returns: state
         """
+        self._ensure_cache(worker_context)
+
+        if self.driver.get_state(worker_context) == states.GONE:
+            self.log.debug('%s driver reported its state is GONE',
+                           self.driver.RESOURCE_NAME)
+            self.state = states.GONE
+            return self.state
+
+        if self.instance_info is None:
+            self.log.info(_LI('no backing instance, marking as down'))
+            self.state = states.DOWN
+            return self.state
+
         for i in xrange(cfg.CONF.max_retries):
-            if akanda_client.is_alive(self.instance_info.management_address,
-                                      cfg.CONF.akanda_mgt_service_port):
+            if self.driver.is_alive(self.instance_info.management_address):
                 if self.state != states.CONFIGURED:
                     self.state = states.UP
                 break
@@ -200,12 +204,13 @@ class InstanceManager(object):
             self._boot_counter.reset()
         return self.state
 
-    @ensure_cache
     def boot(self, worker_context):
         """Boots the instance with driver pre/post boot hooks.
 
         :returns: None
         """
+        self._ensure_cache(worker_context)
+
         self.log.info('Booting %s' % self.driver.RESOURCE_NAME)
         self.state = states.DOWN
         self._boot_counter.start()
@@ -238,13 +243,14 @@ class InstanceManager(object):
             self.instance_info = instance_info
 
         # driver post boot hook
-        self.driver.post_boot()
+        self.driver.post_boot(worker_context)
 
     def check_boot(self, worker_context):
         """Checks status of instance, if ready triggers self.configure
         """
-        if self.update_state(worker_context, silent=True) in states.READY:
-            self.log.info('Router has booted, attempting initial config')
+        state = self.update_state(worker_context, silent=True)
+        if state in states.READY_STATES:
+            self.log.info('Instance has booted, attempting initial config')
             self.configure(worker_context, states.BOOTING, attempts=1)
             if self.state != states.CONFIGURED:
                 self._check_boot_timeout()
@@ -253,7 +259,6 @@ class InstanceManager(object):
         self.log.debug('Instance is %s' % self.state.upper())
         return False
 
-    @ensure_cache
     @synchronize_driver_state
     def set_error(self, worker_context, silent=False):
         """Set the internal and neutron status for the router to states.ERROR.
@@ -263,11 +268,11 @@ class InstanceManager(object):
         supposed to do what it's told and not make decisions about
         whether or not the router is fatally broken.
         """
+        self._ensure_cache(worker_context)
         self.state = states.ERROR
         self.last_error = datetime.utcnow()
         return self.state
 
-    @ensure_cache
     @synchronize_driver_state
     def clear_error(self, worker_context, silent=False):
         """Clear the internal error state.
@@ -279,6 +284,7 @@ class InstanceManager(object):
         """
         # Clear the boot counter.
         self._boot_counter.reset()
+        self._ensure_cache(worker_context)
         self.state = states.DOWN
         return self.state
 
@@ -290,11 +296,10 @@ class InstanceManager(object):
             seconds_since_error = (
                 datetime.utcnow() - self.last_error
             ).total_seconds()
-            if seconds_since_error < cfg.CONF.error_state_coolstates.DOWN:
+            if seconds_since_error < cfg.CONF.error_state_cooldown:
                 return True
         return False
 
-    @ensure_cache
     @synchronize_driver_state
     def stop(self, worker_context):
         """Attempts to destroy the instance with configured timeout.
@@ -302,7 +307,12 @@ class InstanceManager(object):
         :param worker_context:
         :returns:
         """
+        self._ensure_cache(worker_context)
         self.log.info(_LI('Destroying instance'))
+
+        if not self.instance_info:
+            self.log.info(_LI('Instance already destroyed.'))
+            return
 
         try:
             worker_context.nova_client.destroy_instance(self.instance_info)
@@ -312,7 +322,7 @@ class InstanceManager(object):
         start = time.time()
         while time.time() - start < cfg.CONF.boot_timeout:
             if not worker_context.nova_client.\
-                    get_instance_by_id(self.self.instance_info._id):
+                    get_instance_by_id(self.instance_info.id_):
                 if self.state != states.GONE:
                     self.state = states.DOWN
                 return self.state
@@ -322,7 +332,6 @@ class InstanceManager(object):
             'Router failed to stop within %d secs'),
             cfg.CONF.boot_timeout)
 
-    @ensure_cache
     def configure(self, worker_context,
                   failure_state=states.RESTART, attempts=None):
         """Pushes config to instance
@@ -336,12 +345,14 @@ class InstanceManager(object):
         self.state = states.UP
         attempts = attempts or cfg.CONF.max_retries
 
-        interfaces = akanda_client.get_interfaces(
-            self.instance_info.management_address,
-            cfg.CONF.akanda_mgt_service_port
-        )
+        self._ensure_cache(worker_context)
+        if self.driver.get_state(worker_context) == states.GONE:
+            return
 
-        if not self._verify_interfaces(interfaces):
+        interfaces = self.driver.get_interfaces(
+            self.instance_info.management_address)
+
+        if not self._verify_interfaces(self.driver.ports, interfaces):
             # FIXME: Need a states.REPLUG state when we support hot-plugging
             # interfaces.
             self.log.debug("Interfaces aren't plugged as expected.")
@@ -374,11 +385,9 @@ class InstanceManager(object):
 
         for i in xrange(attempts):
             try:
-                akanda_client.update_config(
+                self.driver.update_config(
                     self.instance_info.management_address,
-                    cfg.CONF.akanda_mgt_service_port,
-                    config
-                )
+                    config)
             except Exception:
                 if i == attempts - 1:
                     # Only log the traceback if we encounter it many times.
@@ -403,12 +412,12 @@ class InstanceManager(object):
         :returns:
         """
         self.log.debug('Attempting to replug...')
-        self._ensure_provider_ports(self.driver.details, worker_context)
 
-        interfaces = akanda_client.get_interfaces(
-            self.instance_info.management_address,
-            cfg.CONF.akanda_mgt_service_port
-        )
+        self.driver.pre_plug(worker_context)
+
+        interfaces = self.driver.get_interfaces(
+            self.instance_info.management_address)
+
         actual_macs = set((iface['lladdr'] for iface in interfaces))
         instance_macs = set(p.mac_address for p in self.instance_info.ports)
         instance_macs.add(self.instance_info.management_port.mac_address)
@@ -426,7 +435,7 @@ class InstanceManager(object):
         instance_ports = {p.network_id: p for p in self.instance_info.ports}
         instance_networks = set(instance_ports.keys())
 
-        logical_networks = set(p.network_id for p in self.router_obj.ports)
+        logical_networks = set(p.network_id for p in self.driver.ports)
 
         if logical_networks != instance_networks:
             instance = worker_context.nova_client.get_instance_by_id(
@@ -436,7 +445,7 @@ class InstanceManager(object):
             # For each port that doesn't have a mac address on the instance...
             for network_id in logical_networks - instance_networks:
                 port = worker_context.neutron.create_vrrp_port(
-                    self.router_obj.id,
+                    self.driver.id,
                     network_id
                 )
                 self.log.debug(
@@ -477,19 +486,33 @@ class InstanceManager(object):
             self.log.debug(
                 "Waiting for interface attachments to take effect..."
             )
-            interfaces = akanda_client.get_interfaces(
-                self.instance_info.management_address,
-                cfg.CONF.akanda_mgt_service_port
-            )
-            if self._verify_interfaces(self.driver.details, interfaces):
+            interfaces = self.driver.get_interfaces(
+                self.instance_info.management_address)
+
+            if self._verify_interfaces(self.driver.ports, interfaces):
                 # replugging was successful
                 # TODO(mark) update port states
                 return
+
             time.sleep(1)
             replug_seconds -= 1
 
         self.log.debug("Interfaces aren't plugged as expected, rebooting.")
         self.state = states.RESTART
+
+    def _ensure_cache(self, worker_context):
+        if not self.instance_info:
+            self.instance_info = (
+                worker_context.nova_client.get_instance_info(self.driver.name)
+            )
+
+            if self.instance_info:
+                (
+                    self.instance_info.management_port,
+                    self.instance_info.ports
+                ) = worker_context.neutron.get_ports_for_instance(
+                    self.instance_info.id_
+                )
 
     def _check_boot_timeout(self):
         """If the instance was created more than `boot_timeout` seconds
@@ -518,18 +541,17 @@ class InstanceManager(object):
                 if self.state != states.ERROR:
                     self.state = states.DOWN
 
-    def _verify_interfaces(self, logical_config, interfaces):
+    def _verify_interfaces(self, ports, interfaces):
         """Verifies the network interfaces are what they should be.
         """
         actual_macs = set((iface['lladdr'] for iface in interfaces))
         self.log.debug('MACs found: %s', ', '.join(sorted(actual_macs)))
-
         if not all(
-            getattr(p, 'mac_address', None) for p in logical_config.ports
+            getattr(p, 'mac_address', None) for p in ports
         ):
             return False
 
-        num_logical_ports = len(list(logical_config.ports))
+        num_logical_ports = len(list(ports))
         num_instance_ports = len(list(self.instance_info.ports))
         if num_logical_ports != num_instance_ports:
             return False
@@ -540,13 +562,3 @@ class InstanceManager(object):
         self.log.debug('MACs expected: %s', ', '.join(sorted(expected_macs)))
 
         return actual_macs == expected_macs
-
-    def _ensure_provider_ports(self, instance_obj, worker_context):
-        if instance_obj.external_port is None:
-            # FIXME: Need to do some work to pick the right external
-            # network for a tenant.
-            self.log.debug('Adding external port to instance')
-            ext_port = worker_context.neutron.\
-                create_external_port(instance_obj)
-            instance_obj.external_port = ext_port
-        return instance_obj
